requirements.txt
requests
python-dotenv
pandas
numpy
matplotlib
seaborn
scikit-learn
jupyterlab
hopsworks
hsfs
streamlit
xgboost
joblib

